Refactored the start of the server
